1.19Programming Problem-2
Starting with the Draw program provided in Sect. 20.1, extend the program to
include a new button to draw a new shape for the Draw program. For instance,
have the draw program draw a star on the screen or a smiley face or something
of your choosing. HINT: If you use the forward and back methods to draw your
shape, you can scale it by multiplying each forward and back amount by a scale
value. Then, you can let the user pick a scale for it (or use the radius amount
as your scale) and draw your shape in whatever size you like. To complete this
exercise you must extend your XML format to include a new graphics command
to store the relevant information for drawing your new shape. You must also
define a new graphicsCommand class for your new shape.