Added files via upload
Conduct an experiment to prove that the product of two numbers does not depend
on the size of the two numbers being multiplied. Write a program that plots the
results of multiplying numbers of various sizes together. HINT: To get a good
reading you may want to do more than one of these multiplications and time
them as a group since a multiplication happens pretty quickly in a computer.
Verify that it truly is a O(1) operation. Do you see any anomalies? It might be
explained by Python’s support of large integers. What is the cutoff point for
handling multiplications in constant time? Why? Write a program that produces
an XML file with your results in the format given in this chapter. Then visualize
your results with the PlotData.py program provided in this chapter.